requirements.txt
alembic==1.4.2
aniso8601==8.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
mysqlclient
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
Werkzeug==1.0.1
